requirements.txt
tensorflow==2.12
keras
pillow
numpy
flask